Move Graph to Core
